Made Bot.Control an upper level class 'Control'. Moved contents of Bot into Control. Refactored IBot to Bot.
