+++ src/main/java/com/github/otbproject/otbproject/command/scheduler/ScheduledCommand.java
@@ -1,7 +1,7 @@
 package com.github.otbproject.otbproject.command.scheduler;
 
 import com.github.otbproject.otbproject.App;
-import com.github.otbproject.otbproject.bot.Bot;
+import com.github.otbproject.otbproject.bot.Control;
 import com.github.otbproject.otbproject.channel.Channel;
 import com.github.otbproject.otbproject.messages.receive.PackagedMessage;
 import com.github.otbproject.otbproject.messages.send.MessagePriority;
@@ -14,7 +14,7 @@ public class ScheduledCommand implements Runnable {
 
     public ScheduledCommand(Channel channel, String command) {
         this.channel = channel;
-        packagedMessage = new PackagedMessage(command, Bot.getBot().getUserName(), channel.getName(), channel.getName(), UserLevel.INTERNAL, MessagePriority.DEFAULT);
+        packagedMessage = new PackagedMessage(command, Control.getBot().getUserName(), channel.getName(), channel.getName(), UserLevel.INTERNAL, MessagePriority.DEFAULT);
     }
 
     @Override
