+++ src/main/java/com/github/otbproject/otbproject/util/Watcher.java
@@ -1,7 +1,7 @@
 package com.github.otbproject.otbproject.util;
 
 import com.github.otbproject.otbproject.App;
-import com.github.otbproject.otbproject.bot.Bot;
+import com.github.otbproject.otbproject.bot.Control;
 import com.github.otbproject.otbproject.gui.GuiApplication;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 
@@ -19,7 +19,7 @@ public class Watcher {
                         App.logger.error("Watcher thread crashed");
                         App.logger.error("Please report watcher crash to the developers");
                         App.logger.catching(e);
-                        if (Bot.Graphics.present()) {
+                        if (Control.Graphics.present()) {
                             GuiApplication.errorAlert("Watcher Error", "OTB watcher has encountered an error");
                         }
                     }))
@@ -37,7 +37,7 @@ public class Watcher {
             if ((numExceptions >= EXCEPTION_LIMIT) && (numExceptions < EXCEPTION_LIMIT + DUPLICATE_LIMIT)) {
                 App.logger.error("OTB has experienced multiple internal errors");
                 App.logger.error("Please report this problem to the developers");
-                if (Bot.Graphics.present()) {
+                if (Control.Graphics.present()) {
                     GuiApplication.errorAlert("Internal Error", "OTB has experienced multiple internal errors");
                 }
             }
